Set up content provider after download completion
If the CordovaActivity is still running when
the download finishes, then it won't re-initialize
the plugin, it won't re-initialize the content
provider, and the content provider will be stuck
thinking the expansion file doesn't exist yet,
until the app is force-quit and re-launched.

So, we'll manually send a message to the ContentProvider
telling to to re-initialize as needed.